ESPEED firmware v6.9 (#33)
* feat(scripts): cross-platform OS detection for serial port and Arduino tools

upload_spiffs.sh and flash_all.sh now detect the host OS and adapt:
- arduino_data_dir(): macOS ~/Library/Arduino15, Linux ~/.arduino15,
  Windows (Git Bash) %LOCALAPPDATA%/Arduino15
- find_serial_port(): macOS /dev/cu.usbserial*, Linux /dev/ttyUSB*/ttyACM*,
  Windows: skips auto-detect (COM ports require -p COMx)
- port_hint(): OS-appropriate hint shown on port-not-found errors
- .exe suffix fallback for mkspiffs/esptool on Windows


* feat(firmware): v6.9 — configurable BRAKE and SENSI encoder step size

Add SETTINGS → DISPLAY → BRAKE STEP and SENSI STEP submenus that let the
user control how many percent each encoder tick moves the value while editing.
Brake step: 1–50% per tick (default 1%). Sensi step: 0.5–5.0% in 0.5%
increments (default 0.5%). Both use the same delta-tick approach as the
existing ANTIS step mechanism. Also fixes ANTIS mode label from "MS" to "ms".

Bumps STORED_VAR_VERSION to 23 and SW_MINOR_VERSION to 9.


* docs: add BRAKE STEP and SENSI STEP to menu tree (all languages)

Document the new DISPLAY → BRAKE STEP and DISPLAY → SENSI STEP submenus
in all 7 language docs (en, no, de, es, it, nl, pt) — both the description
bullet points and the full menu tree.


* feat(control): support 0.1% brake and sensi tuning

* fix(ui): scale brake and sensi step editing

* fix(ui): edit brake and sensi steps inline

* build(release): add manual AS5600 assets and clarify OTA path

* Pin AS5600 library to v0.6.7 in CI workflows for reproducible builds

Agent-Logs-Url: https://github.com/nerdweeg/ESPEED32/sessions/ba3081b7-3fcc-4a9e-a9e9-fb7b68fbc22c


* fix(display): correct BRAKE and SENSI step row formatting

* fix(ui): snap car params to step grid on step change or car select

- formatDisplayStepPercentValue always emits one decimal (e.g. "1.0%")
  so the string width is constant while scrolling, preventing residual
  characters on the OLED when crossing whole-percent boundaries.
- snapCarParamToCurrentSteps() floors brake and sensi to the nearest
  multiple of the current step size, eliminating hidden sub-step decimals.
- Called on BRAKE STEP / SENSI STEP confirm in showDisplaySettings().
- Called after showCarSelection() so switching cars also snaps params.


* feat(ui): move ANTISPIN+BRAKE/SENSI STEP into a STEPS submenu

Display settings menu shrinks from 9 to 7 items by collapsing
ANTISPIN, BRAKE STEP and SENSI STEP into a new STEPS submenu.
This reduces top-level clutter and groups all stepping/input
behaviour settings in one place.

- slot_ESC.h: DISPLAY_ITEMS_COUNT 9→7, add STEPS_ITEMS_COUNT 4
- ui_strings: new DISPLAY_MENU_NAMES (7) + STEPS_MENU_NAMES (4)
  in all 9 languages, both UPPER and Pascal variants
- ui_text_access: add getStepsMenuName()
- app_init: update initDisplayMenuItems(), add initStepsMenuItems()
- settings_display_submenus: add showStepsSettings() with full
  screensaver/idle handling, inline BRAKE/SENSI STEP editing,
  snap-on-confirm and ANTISPIN passthrough
- settings_display_menu: replace ANTISPIN inline block with
  STEPS submenu call; remove now-redundant snap code and
  formatDisplayStepPercentValue helper


* fix(build): add missing g_settingsMenu extern in settings_display_submenus.cpp


* docs: update all 7 language manuals for STEPS submenu restructure

ANTISPIN, BRAKE STEP and SENSI STEP moved under a new STEPS submenu.
Updated in all languages:
- Menu tree: wrap the three items under STEPS with correct BACK entry
- Body text: update all path references from DISPLAY->X to
  DISPLAY->STEPS->X
- ANTIS table row: update DISPLAY->ANTISPIN path

Languages updated: en, no, es, de, it, nl, pt


* fix(ui): snap stepped brake and sensi edits safely

* fix(ui): reduce brake and sensi row flicker

* docs(web): regenerate localized manuals

---------

+++ source/ESPEED32/telemetry_logging.h
@@ -45,10 +45,10 @@ typedef struct {
   uint32_t t_ms;
   uint16_t vin_mV;
   uint16_t current_mA;
-  uint16_t sensi_halfPct;
+  uint16_t sensi_halfPct; /* [0.1%] Historical field name kept for storage/API compatibility */
   uint8_t trigger_pct;
   uint8_t output_pct;
-  uint8_t brake_pct;
+  uint16_t brake_pct;      /* [0.1%] Applied brake value */
   uint8_t carIndex;
   uint8_t releaseMode;
   uint8_t flags;
@@ -110,7 +110,7 @@ void telemetryCaptureSample(uint8_t carIndex,
                             uint8_t outputPct,
                             uint16_t vin_mV,
                             uint16_t current_mA,
-                            uint8_t brakePct,
+                            uint16_t brakePct,
                             uint16_t sensi_halfPct,
                             uint8_t releaseMode,
                             uint8_t flags);
